Move cards components to features
